crawler/persist/itemsaver.go
package persist

import (
	"context"
	"log"

	"gopkg.in/olivere/elastic.v5"
)

func ItemSaver() chan interface{} {
	out := make(chan interface{})
	go func() {
		itemCount := 0
		for {
			item := <-out
			log.Printf("itemSaver got item: #%d: %v", itemCount, item)
			itemCount++

			_, err := saver(item)
			if err != nil {
				log.Printf("item saver error: saving item %v: %v",
					item, err)
			}
		}
	}()
	return out
}

func saver(item interface{}) (string, error) {
	client, err := elastic.NewClient(elastic.SetSniff(false))
	if err != nil {
		return "", err
	}
	resp, err := client.Index().
		Index("dating_profile").
		Type("zhenai").BodyJson(item).
		Do(context.Background())
	if err != nil {
		return "", err
	}
	return resp.Id, nil
}
